chatgpt changes,levelpopup odne
