+++ src/components/LevelPopup/LevelPopup.jsx
@@ -18,7 +18,6 @@ const LevelPopup = ({ claimedLevels, onClose }) => {
   return (
     <div>
       {/* <!-- login popup start here --> */}
-      {/* {claimedLevels?.map((badges, index) => { */}
       return (
       <section class="loginPopup silverPopup">
         <div class="wrapper">
@@ -28,10 +27,6 @@ const LevelPopup = ({ claimedLevels, onClose }) => {
               alt="Genaiguru medal-second-place.png"
               title="Genaiguru medal-second-place.png"
             />
-            {/* <div class="dotsImg">
-              <span></span>
-              <span></span>
-            </div> */}
           </figure>
           <div class="labelInner">
             <h6>{claimedLevels.name} </h6>
@@ -49,8 +44,6 @@ const LevelPopup = ({ claimedLevels, onClose }) => {
         </div>
       </section>
       );
-      {/* })} */}
-      {/* <!-- login popup end here -->   */}
     </div>
   );
 };
